Publishing Blog #1
Supported by the Sloan Foundation, we introduce USER, a research initiative that explores how open source scientific and research software teams understand, consider, and undertake usability and design opportunities in their projects. As part of our active work on this project, we are openly sharing our progress along the way in a series of blog posts.